brief commenter

+++ exerciceArticle/Article.cs
@@ -1,11 +1,14 @@
 ﻿using System;
 namespace exerciceArticle
 {
+	/// <summary>
+	/// Classe de l'Article
+	/// </summary>
 	public class Article
 	{
 		public string Nom { get; set; }
 		public int Reference { get; set; }
-		public int Prix { get; set; }
+		public decimal Prix { get; set; }
 
 		public Article()
 		{
@@ -14,6 +17,10 @@ namespace exerciceArticle
 			Prix = 0;
 		}
 
+		/// <summary>
+		/// Affiche le contenu d'un article
+		/// </summary>
+		/// <returns>Chaine de caractère des attributs</returns>
         public override string ToString()
         {
             return $"Nom: {Nom}, Référence: {Reference}, Prix: {Prix}€";
