+++ exerciceArticle/Magasin.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 namespace exerciceArticle
 {
+	/// <summary>
+	/// classe magasin regroupant tous les articles
+	/// </summary>
 	public class Magasin
 	{
 		List<Article> ListesArticles { get; set; }
@@ -10,11 +13,20 @@ namespace exerciceArticle
 			ListesArticles = new();
 		}
 
+		/// <summary>
+		/// Ajoute un article dans la liste
+		/// </summary>
+		/// <param name="article">objet article</param>
 		public void ajouterArticle(Article article)
 		{
 			ListesArticles.Add(article);
 		}
 
+		/// <summary>
+		/// Test si il existe dans la liste un article avec la même référence
+		/// </summary>
+		/// <param name="article">objet article</param>
+		/// <returns>renvoie vrai ou faux</returns>
 		public bool testRef(Article article)
 		{
             foreach (Article article1 in ListesArticles)
@@ -27,6 +39,10 @@ namespace exerciceArticle
             }
 			return false;
         }
+
+		/// <summary>
+		/// Affiche tous les articles
+		/// </summary>
 		public void afficherTousLesArticles()
 		{
 			foreach(Article article in ListesArticles)
@@ -35,6 +51,10 @@ namespace exerciceArticle
 			}
 		}
 
+		/// <summary>
+		/// Affiche un article par référence de l'article
+		/// </summary>
+		/// <param name="reference">la référence en nombre</param>
 		public void afficheParRef(int reference)
 		{
 			foreach(Article article in ListesArticles)
@@ -50,6 +70,10 @@ namespace exerciceArticle
 			}
 		}
 
+		/// <summary>
+		/// Affiche un article par nom de l'article
+		/// </summary>
+		/// <param name="nom">chaine de caractère du nom</param>
 		public void afficherParNom(string nom)
 		{
             foreach (Article article in ListesArticles)
@@ -65,6 +89,11 @@ namespace exerciceArticle
             }
         }
 
+		/// <summary>
+		/// Modifie un article par référence dans la liste
+		/// </summary>
+		/// <param name="reference">la référence en nombre</param>
+		/// <param name="article">objet article</param>
 		public void modifierParRef(int reference, Article article)
 		{
 			for(int index = 0; index < ListesArticles.Count(); index++)
@@ -80,6 +109,10 @@ namespace exerciceArticle
             }
 		}
 
+		/// <summary>
+		/// Suprime l'article par référence
+		/// </summary>
+		/// <param name="reference">la référence en nombre</param>
 		public void supprimerParRef(int reference)
 		{
             for (int index = 0; index < ListesArticles.Count(); index++)
@@ -95,7 +128,12 @@ namespace exerciceArticle
             }
         }
 
-		public void afficherParPrix(int prix1,int prix2)
+		/// <summary>
+		/// Affiche les articles compris entre 2 prix
+		/// </summary>
+		/// <param name="prix1">1er prix en nombre</param>
+		/// <param name="prix2">2ème prix en nombre</param>
+		public void afficherParPrix(decimal prix1,decimal prix2)
 		{
             foreach (Article article in ListesArticles)
             {
